added some directories ready for development.
I decided to nearly rewrite the kernel as i have been running into trouble when getting isrs working despite many attempts to fix it. I will be working mainly with theory from the osdev wiki and will try to make it entirely from scratch minus stuff i have made in the past that i will port to this OS. This is going to be a big project that i will continue working on for a while.